Add babel_config_version setting
This is intended to be a value that translates to the "composite" version
of babel-core and each babel preset and plugin in the consumer's application.
The purpose is to use it to expire the cache for compiled assets. So, when
you upgrade the version of babel-core or you add/remove/upgrade a babel
plugin or preset in your babel options, you'd increment this version which
will cause the Transformer's cache key to change.