Add test_app that transforms ES2015 modules to AMD proving support
This was created by copying the Rails 5.0 test app, replacing shared
files with symlinks, and then making the changes to transform modules
to AMD (which are loaded using almond). This app naturally does not
symlink the asset_compilation_test.rb integration test, since it
expects AMD (not UMD) modules in the transpiled code.