Default the root_dir setting to Rails.root.to_s in Rails applications
While sprockets and sprockets-bumble_d can be used outside of Rails
applications, they are most commonly used in Rails apps. In a Rails
app, it's likely that `node_modules` will be installed in the rails
root, so this default should work out of the box for most people that
use this gem with Rails. Thus, they'll have one less thing to configure.